bin/sbuild-deb-package: change package versioning to mitigate errors during dist-upgrades.
The local version now contains a numerical version of the code name
instead of a string, which is lexicographically compared and will wrap
around at some point for Ubuntu, or just be plain random between
releases for Debian.

Descriptions have been updated to include a capitalized distro name and
also the numerical version in addition to the code name.

In order to make the transition work, ALL packages (releases and
nightlies) must be revbumped. For nightlies, one commit to each repo is
enough to generate a new git date and commit number, while release
packages need a proper revision or version bump. Merely rebuilding with
the old version is not enough, as the new packages will be seen as older
than the old ones.